Update utils.py
It will keep sending requests to Huggingface to verify tokens. If the frequency exceeds the limit, Huggingface will reject it. This results in even if Huggingface is set to write token, it will still prompt that it is not
The following error message will appear:
autotrain.app.utils:token_verification:138 - Failed to request /api/whoami-v2 - 429
ERROR    | 2025-03-24 01:57:16 | autotrain.app.ui_routes:user_authentication:343 - Failed to verify token: Invalid token (/api/whoami-v2). Please login with a write token.
This update is to address this issue, I have added a caching mechanism